flask_project/requirements.txt
Flask==3.0.0
Flask-JWT-Extended==4.6.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
Jinja2==3.1.2
pytz==2023.3
psycopg2-binary==2.9.9
Werkzeug==3.0.1
rich==13.7.0
python-dotenv==1.0.0
pika
Flask-Migrate==4.0.5
requests